Move SyntaxClassifier to a separate module
rdar://98318240
